Move CreateGradientMaskInvocation to its own file. No functional changes.
